// src/app/main/shared/models/Laboratorios.ts
export interface Laboratorios{
  swPresionArterial: boolean;
  presionArterial: string | null;
  swGlucosa: boolean;
  glucosa: string | null;
  swColesterol: boolean;
  colesterol: string | null;
  swTrigliceridos: boolean;
  trigliceridos: string | null;
  swHemoglobina: boolean;
  hemoglobina: string | null;
  swGlicasilada: boolean;
  glicasilada: string | null;
}